refactor: FRONTEND: move singleton components to components/singles
