+++ frontend/src/components/singles/Search.jsx
