+++ frontend/src/components/singles/SessionElement.jsx
